requirements.txt
Flask
Gunicorn
mysql-connector-python==8.1.0
Werkzeug==2.2.3
geopy==2.3.0
